requirements.txt
paramiko
pyyaml
boto3
boto
troposphere
colorama
mock
